replaced projGNCG by projGN (#33)
* - added function to explicitly construct Hessian: getHessian
- added projGN: allows to use different solver
* added documentation 
* updated tests
* removed projGNCG / made small changes and added testing
* documented solveGN

* 1) renamed explicit GN solver
2) added check and use pinv now if system is singular

* debugged and added testing for regularizers in 2D

+++ src/LinearSolvers/mumpsWrapper.jl
@@ -38,7 +38,7 @@ if hasMUMPS
 		return MUMPSsolver(Ainv,doClear,ooc,sym,0,0.0,0,0.)
 	end
 
-	function solveLinearSystem!(A,B,X,param::MUMPSsolver,doTranspose=0)
+	function solveLinearSystem!(A,B,X,param::MUMPSsolver,doTranspose::Int=0)
 		if param.doClear == 1
 			clear!(param)
 		end
@@ -51,7 +51,9 @@ if hasMUMPS
 		end
 
 		tic()
+
 		U = applyMUMPS!(param.Ainv, B,X,doTranspose)
+
 		param.solveTime+=toq()
 		param.nSolve+=1
